Move the provided example DFU mod into code, so it can be better looked at than through Github
Includes an implementation for creating a fix definition for the material set before granite was introduced using the Data Structure provided by Exa. This was used as just a quick test to see if all the needed information could be gathered. It should be considered switching to another data structure if required.